Split up misc tests
